Gaussian noise removal, more blur
Highly effective in removing Gaussian noise from an image. In this technique instead of box filter, a Gaussian kernel is used. Standard deviation in X and Y directions also need to be specified along with height and width of the kernel.
Syntax: cv2.GaussianBlur(src, ksize, sigmaX[, dst[, sigmaY[, borderType]]])