Video capture and Face mapping
Using Open CV and Dlib together to create a face recognition window.
thus we will capture our Video and display it with the help of Open CV.
recognize landmarks on the face with the help of Dlib and shape_predictor_68_face_landmarks.dat which is important while working in a face detection project or project concerning motion or object detection.